Extra perimeters on overhangs vs Bridging fixup: Only build extra overhangs if no bridging is possible.
From my understanding “Extra perimeters on overhangs” detects areas where bridges cannot be anchored, and fills them with extra perimeter paths.

However, the old logic only forced bridging when the unbridgeable area was less than 20% of the overhang area AND unsupported distance was less than 20% of the overhand total length.  Which basically boils down to always creating extra perimeters on overhangs if the unbridgable area is greater than 20%.

This change ensures that we bridge an area if any part of it is bridgable and if it is not bridgable then treat it as an overhang and create the extra perimeters to fill it.

While this change may seem heavy handed, it seems to fix a lot of issues people have been having with “Extra perimeters on overhangs”

Now while a better solution might be to bridge the bridgable parts and create extra overhangs on the non-bridgible parts, while investigating this approach I found that for this to work you would need to be able to apply bridging and then overhangs, but forcing this reorder may not be possible at the core of the slicer (please correct me if im wrong).  Additionally, this approach currently requires much more investigation regarding merging of perimeters with overhangs and bridging.