Extra perimeters on overhangs: better test for if first overhang perimeter is anchored
Unfortunately if the first polyline is exactly aligned with a perimeter of optimized_lower_slices it is considered intersecting.

This fix adds a half extrusion width offset because we want to make sure we are actually intersecting with optimized_lower_slices and are truly anchored.

NOTE:  The offset could probably be just as effective at a lower value.